+++ CHANGES.rst
@@ -15,6 +15,9 @@ New features:
 
 Bug fixes:
 
+- Clean up all ``__init__`` methods of the browser views to avoid unnecessary code execution.
+  [thet]
+
 - Make sure the effects of the robotframework REMOTE_LIBRARY_BUNDLE_FIXTURE
   fixture are not accidentally removed as part of tearing down the
   PLONE_APP_CONTENTTYPES_ROBOT_FIXTURE.